+++ src/Buckle/Repl/BelteRepl.Page.cs
@@ -1,0 +1,12 @@
+
+namespace Repl;
+
+public sealed partial class BelteRepl {
+    /// <summary>
+    /// Indicated to the state what page is being displayed to the user.
+    /// </summary>
+    internal enum Page {
+        Repl,
+        Settings
+    }
+}
